Fix Gemini Code Assist issues
✅ Critical fixes:
- Root layout: Moved html/body tags to src/app/layout.tsx (Next.js App Router compliance)
- Simplified frontend layout to return children only

✅ High priority fixes:
- Removed deprecated hideSourceMaps option from Sentry config

✅ Medium priority fixes:
- Replaced @ts-expect-error with type-safe approach for Vercel analytics
- Added proper TypeScript interfaces for window objects

✅ Build verification:
- All 17 pages successfully generated
- No compilation errors
- Ready for production deployment
